新增 - 資料庫改動 - 20251216084923_ChangeAttendanceFeeAmountToDecimal_Migration_Summary 新增 - 後端 - 取得關帳列表、全部簽到、取得關帳資料、新增或編輯關帳、取得關帳列表 修改 - 後端 - 學生費用出席的部分 - 改成可以有小數點 最後統計在四折五入

+++ DoorDB/TblAttendanceFee.cs
@@ -36,13 +36,13 @@ namespace DoorDB
         /// 單堂學費
         /// </summary>
         [Comment("單堂學費")]
-        public int Amount { get; set; }
+        public decimal Amount { get; set; }
 
         /// <summary>
         /// 單堂增減金額 (正數為增加,負數為減少)
         /// </summary>
         [Comment("單堂增減金額")]
-        public int AdjustmentAmount { get; set; }
+        public decimal AdjustmentAmount { get; set; }
 
         /// <summary>
         /// 原始時數總金額
